+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Modified get_deals to use property_history flag (#64, @seanofahey)
 
+* Added hs_forms_from_a_portal_raw and hs_forms_from_a_portal_tidy (#93, @miahatton)
+
 # hubspot 0.0.1
 
 * Added getter functions for retrieving deals and companies